# go-gin-template-main/{{cookiecutter.project_name}}/docs/swagger.json
{
    "schemes": [
        "https",
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "{{cookiecutter.project_description}}",
        "title": "{{cookiecutter.project_name | slugify | lower}}",
        "contact": {}
    },
    "host": "api-dev.fmobile.kz",
    "basePath": "/{{cookiecutter.project_name | slugify | lower}}",
    "paths": {
        "/api/v1/orders": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Создание заказа",
                "parameters": [
                    {
                        "description": "OrderCreateRequest Data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.OrderCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.Response-schema_OrderResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schema.Response-schema_Empty"
                        }
                    }
                }
            }
        },
        "/api/v1/orders/item": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Получение заказа по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "order_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.Response-schema_OrderResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "schema.Empty": {
            "type": "object"
        },
        "schema.OrderCreateRequest": {
            "type": "object",
            "properties": {
                "channel_id": {
                    "type": "integer"
                },
                "customer": {
                    "$ref": "#/definitions/schema.OrderCustomer"
                },
                "order_comment": {
                    "type": "string"
                }
            }
        },
        "schema.OrderCustomer": {
            "type": "object",
            "required": [
                "phone"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
        "schema.OrderResponse": {
            "type": "object",
            "properties": {
                "buyer_id": {
                    "type": "integer"
                },
                "channel_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "order_comment": {
                    "type": "string"
                },
                "order_id": {
                    "type": "integer"
                },
                "order_session": {
                    "type": "string"
                }
            }
        },
        "schema.Response-schema_Empty": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "result": {
                    "$ref": "#/definitions/schema.Empty"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
        "schema.Response-schema_OrderResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "result": {
                    "$ref": "#/definitions/schema.OrderResponse"
                },
                "status": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}